pyproject.toml
[project]
name = "lucien"
version = "0.1.0"
description = "Mac-first document library builder with AI-powered organization"
authors = [
    {name = "Jeff Miller", email = "github@jeffmlr.net"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
keywords = ["document", "organization", "devonthink", "llm", "macos"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Office Suites",
    "Topic :: System :: Archiving",
]

dependencies = [
    "typer>=0.12.0",
    "rich>=13.7.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "pyyaml>=6.0.1",
    "openai>=1.12.0",
]

[project.optional-dependencies]
# Text extraction dependencies (Milestone 2)
extraction = [
    "docling>=1.0.0",
    "pypdf>=4.0.0",
    "python-magic>=0.4.27",
]

# macOS tags dependencies (Milestone 4)
tags = [
    "xattr>=1.0.0",
    "pyobjc-framework-Cocoa>=10.1",
]

# Development dependencies
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "types-pyyaml>=6.0.0",
]

# All optional dependencies
all = [
    "lucien[extraction,tags,dev]",
]

[project.urls]
Homepage = "https://github.com/jeffmlr/lucien"
Documentation = "https://github.com/jeffmlr/lucien#readme"
Repository = "https://github.com/jeffmlr/lucien"
Issues = "https://github.com/jeffmlr/lucien/issues"

[project.scripts]
lucien = "lucien.cli:main_cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["lucien"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--cov=lucien",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.black]
line-length = 120
target-version = ["py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "docling.*",
    "pypdf.*",
    "xattr.*",
]
ignore_missing_imports = true

[tool.coverage.run]
source = ["lucien"]
omit = [
    "*/tests/*",
    "*/test_*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
